# Optimize RLP serialization