--- a/.gitignore
+++ b/.gitignore
@@ -18,3 +18,6 @@ build/
 *.dll
 *.generated.nim
 nimble.paths
+
+#OS specific files
+**/.DS_Store