--- a/eth.nimble
+++ b/eth.nimble
@@ -34,9 +34,7 @@ let cfg =
   " --skipUserCfg --nimcache:build/nimcache -f" &
   " --warning[ObservableStores]:off -d:nimOldCaseObjects" &
   " -d:chronicles_log_level=TRACE" &
-  " --threads:on -d:debug" &
-  " --excessiveStackTrace:on"
-
+  " --threads:on -d:release"
 
 proc build(args, path, outdir: string) =
   exec nimc & " " & lang & " " & cfg & " " & flags & " " & args &