--- a/eth/rlp/default_writer.nim
+++ b/eth/rlp/default_writer.nim
@@ -16,10 +16,12 @@ func writeCount(writer: var RlpDefaultWriter, count: int, baseMarker: byte) =
   else:
     let lenPrefixBytes = uint64(count).bytesNeeded
 
-    writer.output.add baseMarker + (THRESHOLD_LIST_LEN - 1) + byte(lenPrefixBytes)
-    
+    writer.output.add baseMarker + (THRESHOLD_LIST_LEN - 1) +
+      byte(lenPrefixBytes)
+
     writer.output.setLen(writer.output.len + lenPrefixBytes)
-    writer.output.writeBigEndian(uint64(count), writer.output.len - 1, lenPrefixBytes)
+    writer.output.writeBigEndian(uint64(count), writer.output.len - 1,
+      lenPrefixBytes)
 
 proc maybeClosePendingLists(self: var RlpDefaultWriter) =
   while self.pendingLists.len > 0:
@@ -31,12 +33,11 @@ proc maybeClosePendingLists(self: var RlpDefaultWriter) =
     if self.pendingLists[lastListIdx].remainingItems == 0:
       # A list have been just finished. It was started in `startList`.
       let listStartPos = self.pendingLists[lastListIdx].startPos
-
       self.pendingLists.setLen lastListIdx
 
-      let listLen = self.output.len - listStartPos
-
-      let totalPrefixBytes = if listLen < int(THRESHOLD_LIST_LEN): 1
+      let 
+        listLen = self.output.len - listStartPos
+        totalPrefixBytes = if listLen < int(THRESHOLD_LIST_LEN): 1
                             else: int(uint64(listLen).bytesNeeded) + 1
 
       #Shift the written data to make room for the prefix length
@@ -51,8 +52,11 @@ proc maybeClosePendingLists(self: var RlpDefaultWriter) =
         self.output[listStartPos] = LIST_START_MARKER + byte(listLen)
       else:
         let listLenBytes = totalPrefixBytes - 1
-        self.output[listStartPos] = LEN_PREFIXED_LIST_MARKER + byte(listLenBytes)
-        self.output.writeBigEndian(uint64(listLen), listStartPos + listLenBytes, listLenBytes)
+        self.output[listStartPos] = LEN_PREFIXED_LIST_MARKER +
+          byte(listLenBytes)
+
+        self.output.writeBigEndian(uint64(listLen), 
+          listStartPos + listLenBytes, listLenBytes)
     else:
       # The currently open list is not finished yet. Nothing to do.
       return
@@ -72,7 +76,8 @@ func writeInt*(writer: var RlpDefaultWriter, i: SomeUnsignedInt) =
 
 func appendRawBytes*(self: var RlpDefaultWriter, bytes: openArray[byte]) =
   self.output.setLen(self.output.len + bytes.len)
-  assign(self.output.toOpenArray(self.output.len - bytes.len, self.output.len - 1), bytes)
+  assign(self.output.toOpenArray(
+    self.output.len - bytes.len, self.output.len - 1), bytes)
   self.maybeClosePendingLists()
 
 proc writeBlob*(self: var RlpDefaultWriter, bytes: openArray[byte]) =
@@ -91,12 +96,11 @@ proc startList*(self: var RlpDefaultWriter, listSize: int) =
     self.pendingLists.add((listSize, self.output.len))
 
 template finish*(self: RlpDefaultWriter): seq[byte] =
-  doAssert self.pendingLists.len == 0, "Insufficient number of elements written to a started list"
+  doAssert self.pendingLists.len == 0, 
+    "Insufficient number of elements written to a started list"
   self.output
 
 func clear*(w: var RlpDefaultWriter) =
   # Prepare writer for reuse
   w.pendingLists.setLen(0)
   w.output.setLen(0)
-
-