--- a/eth/rlp/length_writer.nim
+++ b/eth/rlp/length_writer.nim
@@ -1,3 +1,10 @@
+# eth
+# Copyright (c) 2019-2025 Status Research & Development GmbH
+# Licensed and distributed under either of
+#   * MIT license (license terms in the root directory or at https://opensource.org/licenses/MIT).
+#   * Apache v2 license (license terms in the root directory or at https://www.apache.org/licenses/LICENSE-2.0).
+# at your option. This file may not be copied, modified, or distributed except according to those terms.
+
 import
   std/options,
   pkg/results,