## Description

This PR improves upon the memory and time complexity of the RLP serialisation algorithm(`RlpWriter`). The existing implementation of RLP uses the `RlpWriter` that extends a buffer and moves memory whenever required. This wastes a lot of time for large payloads. This PR, firstly, implements a `RlpLengthTracker` that dry runs the serialisation to calculate the full length of the serialised data and the prefixes of lists. Secondly, the `RlpTwoPassWriter` uses the pre-calculated values to assign memory in one shot and avoid moving memory to make space for prefixes.

Additionally, the PR also implements a `RlpHashWriter` that directly updates the sponge construction of keccak256 hash to avoid allocating any memory (few exceptions which are negligible).

## Rationale

Naive implementations of RLP serialisation require moving memory because prefixes of lists can only be calculated after serialising the items within the list. Since complex types such as objects are represented as lists the serialisation of any data structure in Ethereum requires moving the encoded/serialised list items to make space for the prefix. In cases where the encoded buffer is hashed, the encoded list items cannot be *absorbed* until the prefix is calculated.

However, the encoded lengths of base types can be calculated without actually serialising them (or allocating memory). This is exploited to split the computation into two. First to calculate lengths of the base types, and then to actually serialise the data. 

By calculating the lengths of the base types we can find the length of encoded list items and thereby also the length of the prefix (without actually serialising the prefix). Additionally, we also derive the total length of the serialised data from this. This logic is implemented as the `RlpLengthTracker` which records `listLengths`, `prefixLengths` and the `totalLength`.

In the second part of the computation, we actually serialise the data (and the prefixes) without skipping over memory or moving memory positions. This is implemented as the `RlpTwoPassWriter`, "two-pass" because we iterate through the data twice. Because the data is iterated through twice the `RlpTwoPassWriter` can only be used within the `encode` function. In cases where users manually call `append`, the `RlpTwoPaasWriter` would require users to manually run the `RlpLengthTracker` for the first pass and then the second pass. 

`RlpHashWriter` also uses two passes but for the second pass it absorbs the data directly into the sponge construction of keccak instead of allocating memory thereby saving memory (amortized constant usage). And for similar reasons as stated above, a new api method, `encodeHash` is exposed.

## Changelog

* introduces a new method `encodeHash` for constant memory RLP encoding and hashing.
* updates the `encode` method to serialise data in "two passes".
* isolates all writer implementations into individual files with `writer.nim` handling only the dependency injection architecture of the `append` method
* introduces two new tests, `test_hashes.nim` and `test_rlp_profiler.nim`. The first, does a sanity check of the hash writer. The second one, time profiles the new writer implementation to the default implementation.

## Benchmarks

using `--mm=refc` (refc GC)
```
CPU Time [Transaction serialization using two pass writer] 0.000001402s
CPU Time [Block Sequence serialization using two pass writer] 0.000002592s
CPU Time [Block serialization using two pass writer] 0.000007493s
.CPU Time [Transaction serialization using default writer] 0.000001559s
CPU Time [Block Sequence serailization using default writer] 0.000003281s
CPU Time [Block serialization using default writer] 0.000009589s
.CPU Time [Transaction hashing using hash writer] 0.000001810s
CPU Time [Block Sequence hashing using hash writer] 0.000006301s
CPU Time [Block hashing using hash writer] 0.000028100s
.CPU Time [Transaction hashing using default writer and then hash] 0.000002291s
CPU Time [Block Sequence hashing using default writer and then hash] 0.000007532s
CPU Time [Block hashing using defualt writer and the hash] 0.000007491s
```

using `--mm=orc` (orc GC)
```
CPU Time [Transaction serialization using two pass writer] 0.000000901s
CPU Time [Block Sequence serialization using two pass writer] 0.000002089s
CPU Time [Block serialization using two pass writer] 0.000005357s
.CPU Time [Transaction serialization using default writer] 0.000000973s
CPU Time [Block Sequence serailization using default writer] 0.000002513s
CPU Time [Block serialization using default writer] 0.000006726s
.CPU Time [Transaction hashing using hash writer] 0.000001609s
CPU Time [Block Sequence hashing using hash writer] 0.000006161s
CPU Time [Block hashing using hash writer] 0.000026231s
.CPU Time [Transaction hashing using default writer and then hash] 0.000001879s
CPU Time [Block Sequence hashing using default writer and then hash] 0.000006502s
CPU Time [Block hashing using defualt writer and the hash] 0.000006492s
```