@jangko 2025-01-07
Nim generic instantiation cache(in the compiler) is known to have a bug. And by looking into the bug tracker, I believe it still there.
https://github.com/nim-lang/Nim/issues?q=is%3Aissue+is%3Aopen+generic+cache
This is only a hunch. The exact problem is yet to be found.