@chirag-parmar 2025-01-07
refc GC
```
CPU Time [Transaction serialization using two pass writer] 0.000001261s
CPU Time [Block Sequence serialization using two pass writer] 0.000002928s
CPU Time [Block serialization using two pass writer] 0.000006690s
CPU Time [Block header serialization using two pass writer] 0.000000858s

CPU Time [Transaction serialization using default writer] 0.000001452s
CPU Time [Block Sequence serailization using default writer] 0.000003362s
CPU Time [Block serialization using default writer] 0.000009260s
CPU Time [Block header serialization using default writer] 0.000001547s

CPU Time [Transaction hashing using hash writer] 0.000001419s
CPU Time [Block Sequence hashing using hash writer] 0.000006039s
CPU Time [Block hashing using hash writer] 0.000027683s
CPU Time [Block header hashing using hash writer] 0.000004401s

CPU Time [Transaction hashing using default writer and then hash] 0.000002267s
CPU Time [Block Sequence hashing using default writer and then hash] 0.000007470s
CPU Time [Block hashing using defualt writer and the hash] 0.000007329s
CPU Time [Block header hashing using hash writer] 0.000005271s
```

orc GC
```
CPU Time [Transaction serialization using two pass writer] 0.000000780s
CPU Time [Block Sequence serialization using two pass writer] 0.000001807s
CPU Time [Block serialization using two pass writer] 0.000005291s
CPU Time [Block header serialization using two pass writer] 0.000000830s

CPU Time [Transaction serialization using default writer] 0.000000930s
CPU Time [Block Sequence serailization using default writer] 0.000002420s
CPU Time [Block serialization using default writer] 0.000006576s
CPU Time [Block header serialization using default writer] 0.000001354s

CPU Time [Transaction hashing using hash writer] 0.000001478s
CPU Time [Block Sequence hashing using hash writer] 0.000005939s
CPU Time [Block hashing using hash writer] 0.000026274s
CPU Time [Block header hashing using hash writer] 0.000004191s

CPU Time [Transaction hashing using default writer and then hash] 0.000001850s
CPU Time [Block Sequence hashing using default writer and then hash] 0.000006511s
CPU Time [Block hashing using defualt writer and the hash] 0.000006459s
CPU Time [Block header hashing using hash writer] 0.000004640s
```